+++ flutter/example/lib/main.dart
@@ -88,6 +88,7 @@ Future<void> setupSentry(
 
       options.maxRequestBodySize = MaxRequestBodySize.always;
       options.maxResponseBodySize = MaxResponseBodySize.always;
+      options.navigatorKey = navigatorKey;
 
       _isIntegrationTest = isIntegrationTest;
       if (_isIntegrationTest) {